+++ src/Common/ToJsonInterface.php
@@ -4,12 +4,24 @@ namespace andrefelipe\Orchestrate\Common;
 /**
  * An object that can be represented as a JSON string.
  */
-interface ToJsonInterface extends ToArrayInterface
+interface ToJsonInterface extends ToArrayInterface, \JsonSerializable
 {
     /**
      * Get the JSON representation of an object. Uses the toArray output.
+     * Exactly the same as using json_encode($keyvalue_item).
      *
      * @return string
      */
     public function toJson($options = 0, $depth = 512);
+
+    /**
+     * Serializes the object for json_encode
+     *
+     *<code>
+     * echo json_encode($keyvalue_item);
+     *</code>
+     *
+     * @return array
+     */
+    public function jsonSerialize();
 }