+++ src/Common/ToJsonTrait.php
@@ -2,7 +2,7 @@
 namespace andrefelipe\Orchestrate\Common;
 
 /**
- * Trait that implements the toJson method.
+ * Trait that implements the JSON serialization methods.
  * Implementation should also add the ToJsonInterface.
  *
  * @internal
@@ -11,6 +11,11 @@ trait ToJsonTrait
 {
     public function toJson($options = 0, $depth = 512)
     {
-        return json_encode($this->toArray(), $options, $depth);
+        return json_encode($this, $options, $depth);
+    }
+
+    public function jsonSerialize()
+    {
+        return $this->toArray();
     }
 }